+++ pii_recognition/evaluation/pakkr_pipeline_test.py
@@ -3,7 +3,12 @@ from unittest.mock import patch
 
 from pii_recognition.registration.registry import Registry
 
-from .pipeline import get_recogniser, get_tokeniser, load_test_data, reader_registry
+from .pakkr_pipeline import (
+    get_recogniser,
+    get_tokeniser,
+    load_test_data,
+    reader_registry,
+)
 
 
 class RegistryNoConfig:
